[docs] minor code refactoring
