+++ docs/source/_ext/meshroom_doc.py
@@ -1,7 +1,6 @@
 from docutils import nodes
 from docutils.parsers.rst import Directive
-from myst_parser.docutils_ import Parser
-from myst_parser.mdit_to_docutils.base import make_document
+from utils import md_to_docutils
 
 import importlib
 from meshroom.core import desc
@@ -25,28 +24,23 @@ class MeshroomDoc(Directive):
         # Class inherits desc.Node
         if issubclass(node_class, desc.Node):
             node = node_class()
-            parser = Parser()
             # Category
-            doc = make_document(parser_cls=Parser)
-            parser.parse('**Category**: {}'.format(node.category), doc)
+            doc = md_to_docutils('**Category**: {}'.format(node.category))
             result.extend(doc.children)
             # Documentation
-            doc = make_document(parser_cls=Parser)
-            parser.parse(node.documentation, doc)
+            doc = md_to_docutils(node.documentation)
             result.extend(doc.children)
             # Inputs
             text_inputs = '**Inputs**: \n'
             for attr in node.inputs:
                 text_inputs += '- {} ({})\n'.format(attr._name, attr.__class__.__name__)
-            doc = make_document(parser_cls=Parser)
-            parser.parse(text_inputs, doc)
+            doc = md_to_docutils(text_inputs)
             result.extend(doc.children)
             # Outputs
             text_outputs = '**Outputs**: \n'
             for attr in node.outputs:
                 text_outputs += '- {} ({})\n'.format(attr._name, attr.__class__.__name__)
-            doc = make_document(parser_cls=Parser)
-            parser.parse(text_outputs, doc)
+            doc = md_to_docutils(text_outputs)
             result.extend(doc.children)
         return result
